requirements.txt
requests==2.18.4
Pillow==5.2.0
unicornhathd==0.0.3
